Removal of IOTG TSN Reference Software for Apollo Lake I
We'll be continue-ing this branch for future development. Hence,
we are removing all existing contents (in this branch only) and
will repopulate it in the next commit.

The apollolake-i branch will continue to host the older version
of IOTG TSN Reference Software - hereby referred to as:
"IOTG TSN Reference Software for Apollo Lake I".

In the next commit, there will be a brand new set of application
and scripts. Its a complete revamp but you will notice that many
elements and learnings from the older version has been brought over.

Signed-off-by: Wong, Vincent Por Yin <vincent.por.yin.wong@intel.com>