Initial commit of the new IOTG TSN Reference Software
This is the first commit, comprising of over 2 years
of hard work, thanks to the many individuals listed below.

This commit introduces 3 separate C-applications and a lot of
new scripts to enable users to try it out quickly. As with
the older version, this reference software is specific to certain
Intel IOTG platforms only - which currently does not include
Apollo Lake I and the i210 Ethernet controller.

As this is an initial commit, some README and instructions may be
incomplete, and near-future commits will address them.

Co-authored-by: Noor Azura Ahmad Tarmizi <noor.azura.ahmad.tarmizi@intel.com>
Co-authored-by: Khor, Isaac Shi Yan <isaac.shi.yan.khor@intel.com>
Co-authored-by: Singh, Sartaj <sartaj.singh@intel.com>

Signed-off-by: Ong, Boon Leong <boon.leong.ong@intel.com>
Signed-off-by: Wong, Vee Khee <vee.khee.wong@intel.com>
Signed-off-by: Kweh Hock Leong <hock.leong.kweh@intel.com>
Signed-off-by: Weifeng Voon <weifeng.voon@intel.com>
Signed-off-by: Song, Yoong Siang <yoong.siang.song@intel.com>