helpers.sh: Use RUNSH_DEBUG_MODE switch to show ForwardErrors
ForwardErrors are internal values that we calculate to see how many
invalid/empty/rogue packets being rx/sent in the opcua-server loop.
Using OPCUA stack, we observed that this happen time to time.
These packets are not counted towards the latency end (discarded)
This is just an indication for users (more to developer actually).
Hence it should only be seen when RUNSH_DEBUG_MODE="YES".
This debug mode now is put by default to NO, to avoid
confusion on many power-user related debug message.

Signed-off-by: Noor Azura Ahmad Tarmizi <noor.azura.ahmad.tarmizi@intel.com>