shell/vs1b.sh: Usage of napi busy polling for 5.10 XDP tx
Napi busy polling is now used to lessen the XDP latency in 5.10 kernels.

The differences between XDP tx in 5.4 and 5.10 that inspire this:
----------------------------------------------------------------
*Tx design in kernel 5.4:
In ndo_xsk_wakeup(), directly call tx function for pkt transmission.

*Tx design in kernel 5.10:
In ndo_xsk_wakeup(), schedule napi so that tx function in napi
polling will start after ksoftirqd is fired.

Signed-off-by: Noor Azura Ahmad Tarmizi <noor.azura.ahmad.tarmizi@intel.com>