opcua_custom: Added sleep after pub and sub Callback
There is additional latency observed in the system after migration to 5.10.
This involves TGL-H/EHL/ADL-S. However the graph produced for opcua based test
stays flat showing that the deviation is low, the system stays deterministic.

The suspicion lies on the pub and sub callback API in libopen62541-iotg fork
is not yielding correctly.

This commit is the workaround to this problem in ADL. Putting sleep after
pub and sub callback, managed to reduce the latency in ADL.

However, the higher average latencies are still observed on return trip
opcua-based pkt and xdp test on TGL-H.

Further investigation is on-going.

Signed-off-by: Noor Azura Ahmad Tarmizi <noor.azura.ahmad.tarmizi@intel.com>