mount `temp_file_dir` directory to tmpfs (#165)
When the `temp_file_dir` directory is not mounted under tmpfs then
it affects the test result on the tsn ref sw test cases.

By solving the issue, we will check on the `temp_file_dir` folder,
if it is not mounted under tmpfs then we will create and mount it.

Signed-off-by: Choong Yong Liang <yong.liang.choong@intel.com>